♻️ refactor: comment API 기존 board.ts에서 분리
